driver-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
odmantic
motor
passlib[bcrypt]
pyjwt
prometheus-client
aio-pika
aioredis